changed EntityInArchetype memory footprint

+++ src/entity_in_archetype.rs
@@ -1,5 +1,5 @@
 #[derive(Clone, Copy)]
 pub(crate) struct EntityInArchetype {
-    pub page_index: usize,
-    pub index_in_page: usize,
+    pub page_index: u32,
+    pub index_in_page: u32,
 }
